# Optimize join operation