Previous implementation of join operation using nested strategy, hence it was really slow when joining big table. In this PR we optimize the join operations. The optimization is done by 2 changes:
* Using hash instead of nested algorithm
* Remove repeatedly append series rows which is slow

This optimization applied to
* INNER JOIN
* LEFT JOIN
* RIGHT JOIN
* OUTER JOIN

Below are the benchmark result of the optimization
```
Running tool: /usr/local/bin/go test -benchmem -run=^$ -bench ^BenchmarkDataFrame_InnerJoin$ github.com/go-gota/gota/dataframe

goos: darwin
goarch: amd64
pkg: github.com/go-gota/gota/dataframe
cpu: Intel(R) Core(TM) i9-9880H CPU @ 2.30GHz
BenchmarkDataFrame_InnerJoin/10_rows_-_10_rows,_1_to_1_join,_2_rows,_random-16         	   87997	     12784 ns/op	   10447 B/op	     118 allocs/op
BenchmarkDataFrame_InnerJoin/10_rows_-_10_rows,_1_to_2_join,_2_rows,_random-16         	   53110	     23191 ns/op	   16789 B/op	     352 allocs/op
BenchmarkDataFrame_InnerJoin/50_rows_-_50_rows,_1_to_1_join,_random-16                 	    6856	    179018 ns/op	  112816 B/op	    2795 allocs/op
BenchmarkDataFrame_InnerJoin/50_rows_-_50_rows,_1_to_1_join,_2_join_rows,_random-16    	   15258	     76330 ns/op	   12177 B/op	     208 allocs/op
BenchmarkDataFrame_InnerJoin/100_rows_-_100_rows,_1_to_1_join,_random-16               	    3033	    413071 ns/op	  217321 B/op	    5362 allocs/op
BenchmarkDataFrame_InnerJoin/300_rows_-_300_rows,_1_to_1_join,_random-16               	     565	   2148514 ns/op	  696389 B/op	   15596 allocs/op
BenchmarkDataFrame_InnerJoin/300_rows_-_300_rows,_1_to_1_join,_2_join_rows,_random-16  	     560	   2236624 ns/op	   14196 B/op	     280 allocs/op
BenchmarkDataFrame_InnerJoin/900_rows_-_900_rows,_1_to_1_join,_2_join_rows,_random-16  	      61	  19667496 ns/op	   10502 B/op	     118 allocs/op
BenchmarkDataFrame_InnerJoin/1000_rows_-_1000_rows,_1_to_1_join,_2_join_rows,_random-16         	      51	  24299993 ns/op	   14201 B/op	     280 allocs/op
PASS
ok  	github.com/go-gota/gota/dataframe	14.046s

Running tool: /usr/local/bin/go test -benchmem -run=^$ -bench ^BenchmarkDataFrame_InnerJoinOptimized$ github.com/go-gota/gota/dataframe

goos: darwin
goarch: amd64
pkg: github.com/go-gota/gota/dataframe
cpu: Intel(R) Core(TM) i9-9880H CPU @ 2.30GHz
BenchmarkDataFrame_InnerJoinOptimized/10_rows_-_10_rows,_1_to_1_join,_2_rows,_random-16         	   63050	     19157 ns/op	   14673 B/op	     221 allocs/op
BenchmarkDataFrame_InnerJoinOptimized/10_rows_-_10_rows,_1_to_2_join,_2_rows,_random-16         	   63591	     19565 ns/op	   14672 B/op	     219 allocs/op
BenchmarkDataFrame_InnerJoinOptimized/50_rows_-_50_rows,_1_to_1_join,_random-16                 	   13558	     89653 ns/op	   99027 B/op	     521 allocs/op
BenchmarkDataFrame_InnerJoinOptimized/50_rows_-_50_rows,_1_to_1_join,_2_join_rows,_random-16    	   26630	     43406 ns/op	   21845 B/op	     545 allocs/op
BenchmarkDataFrame_InnerJoinOptimized/100_rows_-_100_rows,_1_to_1_join,_random-16               	    7996	    154735 ns/op	  187868 B/op	     742 allocs/op
BenchmarkDataFrame_InnerJoinOptimized/300_rows_-_300_rows,_1_to_1_join,_random-16               	    2434	    428118 ns/op	  599601 B/op	    1671 allocs/op
BenchmarkDataFrame_InnerJoinOptimized/300_rows_-_300_rows,_1_to_1_join,_2_join_rows,_random-16  	    5998	    202443 ns/op	   86856 B/op	    2805 allocs/op
BenchmarkDataFrame_InnerJoinOptimized/900_rows_-_900_rows,_1_to_1_join,_2_join_rows,_random-16  	    1970	    650374 ns/op	  311440 B/op	   10039 allocs/op
BenchmarkDataFrame_InnerJoinOptimized/1000_rows_-_1000_rows,_1_to_1_join,_2_join_rows,_random-16         	    1760	    705337 ns/op	  320824 B/op	   11149 allocs/op
PASS
ok  	github.com/go-gota/gota/dataframe	12.838s


Running tool: /usr/local/bin/go test -benchmem -run=^$ -bench ^BenchmarkDataFrame_LeftJoin$ github.com/go-gota/gota/dataframe

goos: darwin
goarch: amd64
pkg: github.com/go-gota/gota/dataframe
cpu: Intel(R) Core(TM) i9-9880H CPU @ 2.30GHz
BenchmarkDataFrame_LeftJoin/10_rows_-_10_rows,_1_to_1_join,_2_rows,_random-16         	   28465	     41820 ns/op	   33791 B/op	     766 allocs/op
BenchmarkDataFrame_LeftJoin/10_rows_-_10_rows,_1_to_2_join,_2_rows,_random-16         	   28773	     41945 ns/op	   33791 B/op	     766 allocs/op
BenchmarkDataFrame_LeftJoin/50_rows_-_50_rows,_1_to_1_join,_random-16                 	    5584	    185805 ns/op	  112816 B/op	    2795 allocs/op
BenchmarkDataFrame_LeftJoin/50_rows_-_50_rows,_1_to_1_join,_2_join_rows,_random-16    	    5728	    190079 ns/op	  120820 B/op	    2962 allocs/op
BenchmarkDataFrame_LeftJoin/100_rows_-_100_rows,_1_to_1_join,_random-16               	    2793	    414948 ns/op	  217320 B/op	    5362 allocs/op
BenchmarkDataFrame_LeftJoin/300_rows_-_300_rows,_1_to_1_join,_random-16               	     558	   2221789 ns/op	  696416 B/op	   15596 allocs/op
BenchmarkDataFrame_LeftJoin/300_rows_-_300_rows,_1_to_1_join,_2_join_rows,_random-16  	     408	   2954309 ns/op	  738711 B/op	   16516 allocs/op
BenchmarkDataFrame_LeftJoin/900_rows_-_900_rows,_1_to_1_join,_2_join_rows,_random-16  	      51	  21787678 ns/op	 2251347 B/op	   48954 allocs/op
BenchmarkDataFrame_LeftJoin/1000_rows_-_1000_rows,_1_to_1_join,_2_join_rows,_random-16         	      45	  26583528 ns/op	 2366813 B/op	   54354 allocs/op
PASS
ok  	github.com/go-gota/gota/dataframe	11.982s

Running tool: /usr/local/bin/go test -benchmem -run=^$ -bench ^BenchmarkDataFrame_LeftJoinOptimized$ github.com/go-gota/gota/dataframe

goos: darwin
goarch: amd64
pkg: github.com/go-gota/gota/dataframe
cpu: Intel(R) Core(TM) i9-9880H CPU @ 2.30GHz
BenchmarkDataFrame_LeftJoinOptimized/10_rows_-_10_rows,_1_to_1_join,_2_rows,_random-16         	   32270	     36523 ns/op	   32253 B/op	     365 allocs/op
BenchmarkDataFrame_LeftJoinOptimized/10_rows_-_10_rows,_1_to_2_join,_2_rows,_random-16         	   31866	     36364 ns/op	   32246 B/op	     363 allocs/op
BenchmarkDataFrame_LeftJoinOptimized/50_rows_-_50_rows,_1_to_1_join,_random-16                 	   12990	     91541 ns/op	   99025 B/op	     521 allocs/op
BenchmarkDataFrame_LeftJoinOptimized/50_rows_-_50_rows,_1_to_1_join,_2_join_rows,_random-16    	   10000	    100047 ns/op	  106872 B/op	     725 allocs/op
BenchmarkDataFrame_LeftJoinOptimized/100_rows_-_100_rows,_1_to_1_join,_random-16               	    7646	    153699 ns/op	  187863 B/op	     742 allocs/op
BenchmarkDataFrame_LeftJoinOptimized/300_rows_-_300_rows,_1_to_1_join,_random-16               	    2722	    427273 ns/op	  599595 B/op	    1671 allocs/op
BenchmarkDataFrame_LeftJoinOptimized/300_rows_-_300_rows,_1_to_1_join,_2_join_rows,_random-16  	    2402	    484306 ns/op	  643338 B/op	    2968 allocs/op
BenchmarkDataFrame_LeftJoinOptimized/900_rows_-_900_rows,_1_to_1_join,_2_join_rows,_random-16  	     820	   1495797 ns/op	 2102538 B/op	   10221 allocs/op
BenchmarkDataFrame_LeftJoinOptimized/1000_rows_-_1000_rows,_1_to_1_join,_2_join_rows,_random-16         	     763	   1596909 ns/op	 2114567 B/op	   11421 allocs/op
PASS
ok  	github.com/go-gota/gota/dataframe	12.640s

Running tool: /usr/local/bin/go test -benchmem -run=^$ -bench ^BenchmarkDataFrame_RightJoin$ github.com/go-gota/gota/dataframe

goos: darwin
goarch: amd64
pkg: github.com/go-gota/gota/dataframe
cpu: Intel(R) Core(TM) i9-9880H CPU @ 2.30GHz
BenchmarkDataFrame_RightJoin/10_rows_-_10_rows,_1_to_1_join,_2_rows,_random-16         	   27195	     44816 ns/op	   34037 B/op	     771 allocs/op
BenchmarkDataFrame_RightJoin/10_rows_-_10_rows,_1_to_2_join,_2_rows,_random-16         	   27736	     43584 ns/op	   33960 B/op	     772 allocs/op
BenchmarkDataFrame_RightJoin/50_rows_-_50_rows,_1_to_1_join,_random-16                 	    6448	    177320 ns/op	  114848 B/op	    2802 allocs/op
BenchmarkDataFrame_RightJoin/50_rows_-_50_rows,_1_to_1_join,_2_join_rows,_random-16    	    5624	    214659 ns/op	  121856 B/op	    2970 allocs/op
BenchmarkDataFrame_RightJoin/100_rows_-_100_rows,_1_to_1_join,_random-16               	    2798	    438853 ns/op	  221417 B/op	    5370 allocs/op
BenchmarkDataFrame_RightJoin/300_rows_-_300_rows,_1_to_1_join,_random-16               	     543	   2251131 ns/op	  712766 B/op	   15606 allocs/op
BenchmarkDataFrame_RightJoin/300_rows_-_300_rows,_1_to_1_join,_2_join_rows,_random-16  	     399	   3041994 ns/op	  747027 B/op	   16529 allocs/op
BenchmarkDataFrame_RightJoin/900_rows_-_900_rows,_1_to_1_join,_2_join_rows,_random-16  	      55	  22517182 ns/op	 2276722 B/op	   48968 allocs/op
BenchmarkDataFrame_RightJoin/1000_rows_-_1000_rows,_1_to_1_join,_2_join_rows,_random-16         	      44	  27815645 ns/op	 2391833 B/op	   54367 allocs/op
PASS
ok  	github.com/go-gota/gota/dataframe	12.495s

Running tool: /usr/local/bin/go test -benchmem -run=^$ -bench ^BenchmarkDataFrame_RightJoinOptimized$ github.com/go-gota/gota/dataframe

goos: darwin
goarch: amd64
pkg: github.com/go-gota/gota/dataframe
cpu: Intel(R) Core(TM) i9-9880H CPU @ 2.30GHz
BenchmarkDataFrame_RightJoinOptimized/10_rows_-_10_rows,_1_to_1_join,_2_rows,_random-16         	   28920	     46395 ns/op	   36309 B/op	     384 allocs/op
BenchmarkDataFrame_RightJoinOptimized/10_rows_-_10_rows,_1_to_2_join,_2_rows,_random-16         	   24666	     44219 ns/op	   33709 B/op	     418 allocs/op
BenchmarkDataFrame_RightJoinOptimized/50_rows_-_50_rows,_1_to_1_join,_random-16                 	   13364	     93748 ns/op	  100199 B/op	     522 allocs/op
BenchmarkDataFrame_RightJoinOptimized/50_rows_-_50_rows,_1_to_1_join,_2_join_rows,_random-16    	    9750	    113959 ns/op	  125051 B/op	     780 allocs/op
BenchmarkDataFrame_RightJoinOptimized/100_rows_-_100_rows,_1_to_1_join,_random-16               	    8072	    161817 ns/op	  189040 B/op	     743 allocs/op
BenchmarkDataFrame_RightJoinOptimized/300_rows_-_300_rows,_1_to_1_join,_random-16               	    2612	    457870 ns/op	  600780 B/op	    1672 allocs/op
BenchmarkDataFrame_RightJoinOptimized/300_rows_-_300_rows,_1_to_1_join,_2_join_rows,_random-16  	    1998	    594861 ns/op	  732893 B/op	    3023 allocs/op
BenchmarkDataFrame_RightJoinOptimized/900_rows_-_900_rows,_1_to_1_join,_2_join_rows,_random-16  	     685	   1778235 ns/op	 2400032 B/op	   10276 allocs/op
BenchmarkDataFrame_RightJoinOptimized/1000_rows_-_1000_rows,_1_to_1_join,_2_join_rows,_random-16         	     649	   1864717 ns/op	 2411449 B/op	   11458 allocs/op
PASS
ok  	github.com/go-gota/gota/dataframe	13.283s

Running tool: /usr/local/bin/go test -benchmem -run=^$ -bench ^BenchmarkDataFrame_OuterJoin$ github.com/go-gota/gota/dataframe

goos: darwin
goarch: amd64
pkg: github.com/go-gota/gota/dataframe
cpu: Intel(R) Core(TM) i9-9880H CPU @ 2.30GHz
BenchmarkDataFrame_OuterJoin/10_rows_-_10_rows,_1_to_1_join,_2_rows,_random-16         	   16772	     71002 ns/op	   57423 B/op	    1324 allocs/op
BenchmarkDataFrame_OuterJoin/10_rows_-_10_rows,_1_to_2_join,_2_rows,_random-16         	   17258	     69843 ns/op	   57425 B/op	    1324 allocs/op
BenchmarkDataFrame_OuterJoin/50_rows_-_50_rows,_1_to_1_join,_random-16                 	    5617	    239613 ns/op	  112816 B/op	    2795 allocs/op
BenchmarkDataFrame_OuterJoin/50_rows_-_50_rows,_1_to_1_join,_2_join_rows,_random-16    	    2840	    382173 ns/op	  229186 B/op	    5572 allocs/op
BenchmarkDataFrame_OuterJoin/100_rows_-_100_rows,_1_to_1_join,_random-16               	    1971	    587741 ns/op	  217322 B/op	    5362 allocs/op
BenchmarkDataFrame_OuterJoin/300_rows_-_300_rows,_1_to_1_join,_random-16               	     322	   3618047 ns/op	  696409 B/op	   15596 allocs/op
BenchmarkDataFrame_OuterJoin/300_rows_-_300_rows,_1_to_1_join,_2_join_rows,_random-16  	     178	   5963814 ns/op	 1416585 B/op	   32735 allocs/op
BenchmarkDataFrame_OuterJoin/900_rows_-_900_rows,_1_to_1_join,_2_join_rows,_random-16  	      27	  44466751 ns/op	 4837021 B/op	   97592 allocs/op
BenchmarkDataFrame_OuterJoin/1000_rows_-_1000_rows,_1_to_1_join,_2_join_rows,_random-16         	      22	  54494459 ns/op	 5067456 B/op	  108392 allocs/op
PASS
ok  	github.com/go-gota/gota/dataframe	13.358s

Running tool: /usr/local/bin/go test -benchmem -run=^$ -bench ^BenchmarkDataFrame_OuterJoinOptimized$ github.com/go-gota/gota/dataframe

goos: darwin
goarch: amd64
pkg: github.com/go-gota/gota/dataframe
cpu: Intel(R) Core(TM) i9-9880H CPU @ 2.30GHz
BenchmarkDataFrame_OuterJoinOptimized/10_rows_-_10_rows,_1_to_1_join,_2_rows,_random-16         	   20296	     57646 ns/op	   52054 B/op	     440 allocs/op
BenchmarkDataFrame_OuterJoinOptimized/10_rows_-_10_rows,_1_to_2_join,_2_rows,_random-16         	   21481	     56722 ns/op	   52045 B/op	     436 allocs/op
BenchmarkDataFrame_OuterJoinOptimized/50_rows_-_50_rows,_1_to_1_join,_random-16                 	   10000	    112319 ns/op	  107449 B/op	     761 allocs/op
BenchmarkDataFrame_OuterJoinOptimized/50_rows_-_50_rows,_1_to_1_join,_2_join_rows,_random-16    	    5932	    218852 ns/op	  202601 B/op	    1165 allocs/op
BenchmarkDataFrame_OuterJoinOptimized/100_rows_-_100_rows,_1_to_1_join,_random-16               	    5584	    219878 ns/op	  205372 B/op	    1234 allocs/op
BenchmarkDataFrame_OuterJoinOptimized/300_rows_-_300_rows,_1_to_1_join,_random-16               	    1978	    583060 ns/op	  666228 B/op	    3259 allocs/op
BenchmarkDataFrame_OuterJoinOptimized/300_rows_-_300_rows,_1_to_1_join,_2_join_rows,_random-16  	    1197	   1044144 ns/op	 1233145 B/op	    5847 allocs/op
BenchmarkDataFrame_OuterJoinOptimized/900_rows_-_900_rows,_1_to_1_join,_2_join_rows,_random-16  	     352	   3878439 ns/op	 4567483 B/op	   20935 allocs/op
BenchmarkDataFrame_OuterJoinOptimized/1000_rows_-_1000_rows,_1_to_1_join,_2_join_rows,_random-16         	     325	   3817070 ns/op	 4594584 B/op	   23436 allocs/op
PASS
ok  	github.com/go-gota/gota/dataframe	13.536s



# Summary
For 10x10 rows the previous implementation and the optimized one having comparable performance, but when we use 50 rows or above the optimized implementation way faster compare to the previous.
For 50x50 rows the optimized on almost 2x faster, and it is even better when number of rows increase significantly, for 1000x1000 row the optimized implementation is almost 17x faster
```