@ariefrahmansyah 2023-01-10
I don't see any benchmark for OuterJoinOptimized. Can you include it also on PR's description?